Move .tina -> tina
This is a recent change to Tina.
