+++ .graphqlrc.yml
@@ -1,1 +1,1 @@
-schema: '.tina/__generated__/schema.gql'
+schema: 'tina/__generated__/schema.gql'
