+++ .prettierignore
@@ -1,1 +1,1 @@
-.tina/__generated__
+tina/__generated__
