+++ package.json
@@ -29,14 +29,14 @@
     "react-icons": "^4.8.0",
     "react-transition-group": "^4.4.5",
     "rooks": "^7.4.3",
-    "sharp": "0.31.3",
+    "sharp": "^0.32.1",
     "signale": "^1.4.0",
     "sitemap": "^7.1.1",
     "sitemapper": "^3.2.6",
     "smartypants": "^0.1.6",
     "spectacle": "^4.0.5",
     "tailwindcss": "^3.2.7",
-    "tinacms": "^1.2.0",
+    "tinacms": "^1.5.2",
     "typeit": "^8.7.1",
     "yargs": "^17.6.2"
   },
@@ -51,7 +51,7 @@
     "@storybook/manager-webpack5": "^6.5.16",
     "@storybook/react": "^6.5.16",
     "@testing-library/react": "^14.0.0",
-    "@tinacms/cli": "^1.3.3",
+    "@tinacms/cli": "^1.5.9",
     "autoprefixer": "^10.4.13",
     "axios": "^1.3.4",
     "babel-eslint": "^10.1.0",
@@ -90,10 +90,10 @@
     "dev:storybook": "start-storybook -p 6006",
     "build": "./bin/cli.js build && tinacms build && next build",
     "build:storybook": "build-storybook",
-    "fix": "prettier --write \"{src,.storybook,.tina,bin}/**/*.{js,ts,md}\" && eslint --fix \"{src,.storybook,.tina,bin}/**/*.{js,ts}\"",
+    "fix": "prettier --write \"{src,.storybook,tina,bin}/**/*.{js,ts,md}\" && eslint --fix \"{src,.storybook,tina,bin}/**/*.{js,ts}\"",
     "test": "npm run test:check && npm run test:lint && npm run test:unit",
-    "test:check": "prettier --check \"{src,.storybook,.tina,bin}/**/*.{js,ts,md}\"",
-    "test:lint": "eslint \"{src,.storybook,.tina,bin}/**/*.{js,ts}\" --max-warnings 0",
+    "test:check": "prettier --check \"{src,.storybook,tina,bin}/**/*.{js,ts,md}\"",
+    "test:lint": "eslint \"{src,.storybook,tina,bin}/**/*.{js,ts}\" --max-warnings 0",
     "test:unit": "jest --passWithNoTests"
   },
   "overrides": {
